installed dependencies for thunk, react-redux and logger. Created Redux store and used thunk middleware. Working on axios call with thunk middleware in the reducers. Updated readme with current API info, removed extra code from file.
