+++ src/index.js
@@ -3,9 +3,19 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 
+import { applyMiddleware, createStore } from 'redux';
+import { Provider } from 'react-redux';
+import logger from 'redux-logger';
+import thunk from 'redux-thunk';
+import { rootReducer } from './common/reducers';
+
+const store = createStore(rootReducer, applyMiddleware(logger, thunk));
+
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Provider store={store}>
+     <App />
+     </Provider>
   </React.StrictMode>,
   document.getElementById('root')
 );
